Move tests for WheelBuilder and friends
